Using (long long) shuts GCC on Linux up.
Whoever defined time_t so vaguely should be embarrassed. It's a portability
nightmare.

+++ format.c
@@ -187,11 +187,11 @@ void format_like_time(Conf *conf)
         timeradd(&sys,  &cmd->rusages[0]->ru_stime, &sys);
     }
 	fprintf(stderr, "real %9lld.%02lld\n",
-      (int64_t) (real.tv_sec / conf->num_runs), (int64_t) ((real.tv_usec / 10000) / conf->num_runs));
+      (long long) (real.tv_sec / conf->num_runs), (long long) ((real.tv_usec / 10000) / conf->num_runs));
 	fprintf(stderr, "user %9lld.%02lld\n",
-      (int64_t) (user.tv_sec / conf->num_runs), (int64_t) ((user.tv_usec / 10000) / conf->num_runs));
+      (long long) (user.tv_sec / conf->num_runs), (long long) ((user.tv_usec / 10000) / conf->num_runs));
 	fprintf(stderr, "sys  %9lld.%02lld\n",
-      (int64_t) (sys.tv_sec / conf->num_runs), (int64_t) ((sys.tv_usec / 10000) / conf->num_runs));
+      (long long) (sys.tv_sec / conf->num_runs), (long long) ((sys.tv_usec / 10000) / conf->num_runs));
 }
